refactor(types): split innerworld.ts into 3 per-entity files
